move graph_to_img to plotting.py
