+++ src/ophys_etl/modules/segmentation/graph_utils/plotting.py
@@ -111,3 +111,35 @@ def create_roi_plot(plot_path: pathlib.Path,
     fig.tight_layout()
     fig.savefig(plot_path)
     return None
+
+
+def graph_to_img(graph_path: pathlib.Path,
+                 attribute: str = 'filtered_hnc_Gaussian') -> np.ndarray:
+    """
+    Convert a graph into a np.ndarray image
+
+    Parameters
+    ----------
+    graph_path: pathlib.Path
+        Path to graph pickle file
+
+    attribute: str
+        Name of the attribute used to create the image
+        (default = 'filtered_hnc_Gaussian')
+
+    Returns
+    -------
+    np.ndarray
+        An image in which the value of each pixel is the
+        sum of the edge weights connected to that node in
+        the graph.
+    """
+    graph = nx.read_gpickle(graph_path)
+    node_coords = np.array(graph.nodes).T
+    row_max = node_coords[0].max()
+    col_max = node_coords[1].max()
+    img = np.zeros((row_max+1, col_max+1), dtype=float)
+    for node in graph.nodes:
+        vals = [graph[node][i][attribute] for i in graph.neighbors(node)]
+        img[node[0], node[1]] = np.sum(vals)
+    return img
